Add recipe dir with file and add link on index.html
I created a new directory called "recipes" to contain all the html files
that will contain the recipes. I added the island-style-fried-rice.html
to this directory. I added a header for this file but nothing else. Not
sure if I'll need to add the boilerplate to this, but it currently
doesn't have a title showing up, so probably does need it, but TOP said
"only add h1" so that's what I'm doing for now. Could be a mistake on
their part, but can always change it later if need be.

Should also have committed sooner. The heading of this commit is too
long. I should've done a commit for:
1) Add recipe dir
2) Add recipe recipe html file
3) Add h1 to recipe html file
4) Add link for recipe to index.html